set up worker logging
